requirements.txt
Django==4.1
mysqlclient==2.1.1
django-money==3.0.0
django-phonenumber-field[phonenumberslite]==7.0.1
pyyaml==6.0
